BTMS/requirements/dev.txt
-r common.txt
coverage==5.5
pep8==1.7.1
flake8==3.7.7
mock==4.0.3
pytest-xdist==2.4.0
pytest-cov==3.0.0
pytest-django==4.4.0
django-environ==0.8.1
